Title: 합과 평균, Time: 126ms, Memory: 10MB, Status: Passed - CodeTree

+++ 231125/합과 평균/sum-and-avg.java
@@ -6,7 +6,7 @@ public class Main {
         int a = in.nextInt();
         int b = in.nextInt();
         int sum = a+b;
-        double average = sum/2;
+        double average = ((double)sum)/2;
         System.out.printf(sum + " " + "%.1f", average);
     }
 }
